UHF-11210: Fix tests
