+++ tests/src/Kernel/AtvServiceHeaderTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Drupal\Tests\helfi_atv\Kernel;
 
 use Drupal\file\Entity\File;
-use Drupal\KernelTests\KernelTestBase;
 use GuzzleHttp\Psr7\Response;
 
 /**
@@ -17,48 +16,12 @@ use GuzzleHttp\Psr7\Response;
  * @covers \Drupal\helfi_atv\AtvService
  * @group helfi_atv
  */
-class AtvServiceHeaderTest extends KernelTestBase {
-  /**
-   * The modules to load to run the test.
-   *
-   * @var array
-   */
-  protected static $modules = [
-    // Drupal.
-    'file',
-    'user',
-    // Contrib.
-    'openid_connect',
-    // Helfi modules.
-    'helfi_api_base',
-    'helfi_atv',
-    'helfi_atv_test',
-    'helfi_helsinki_profiili',
-    // Helsinki profiili requires audit log unnecessarily.
-    'helfi_audit_log',
-  ];
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function setUp(): void {
-    parent::setUp();
-    $this->installConfig(['helfi_atv']);
-    putenv('ATV_API_KEY=fake');
-    putenv('ATV_USE_TOKEN_AUTH=true');
-    putenv('ATV_TOKEN_NAME=tokenName');
-    putenv('ATV_BASE_URL=127.0.0.1');
-    putenv('ATV_VERSION=1.1');
-    putenv('ATV_USE_CACHE=false');
-    putenv('APP_ENV=UNIT_TEST');
-    putenv('ATV_SERVICE=service');
-    putenv('ATV_MAX_PAGES=10');
-  }
+class AtvServiceHeaderTest extends AtvKernelTestBase {
 
   /**
    * Assert if we have correct authorization header.
    */
-  public function assertTokenHeaders($headers): void {
+  private function assertTokenHeaders($headers): void {
     $this->assertArrayHasKey('Authorization', $headers);
     $this->assertArrayNotHasKey('X-Api-Key', $headers);
     $expectedToken = 'Bearer tokenFromMockHelsinkiProfiiliUserData';
@@ -68,7 +31,7 @@ class AtvServiceHeaderTest extends KernelTestBase {
   /**
    * Assert if we have correct API Key header.
    */
-  public function assertApiKeyHeaders($headers): void {
+  private function assertApiKeyHeaders($headers): void {
     $this->assertArrayHasKey('X-Api-Key', $headers);
     $this->assertArrayNotHasKey('Authorization', $headers);
     $expectedKey = getenv('ATV_API_KEY');
