+++ tests/src/Kernel/RequestTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Drupal\Tests\helfi_atv\Kernel;
 
-use Drupal\KernelTests\KernelTestBase;
 use GuzzleHttp\Psr7\Response;
 use League\OpenAPIValidation\PSR7\RequestValidator;
 use League\OpenAPIValidation\PSR7\ValidatorBuilder;
@@ -15,8 +14,7 @@ use League\OpenAPIValidation\PSR7\ValidatorBuilder;
  * @covers \Drupal\helfi_atv\AtvService
  * @group helfi_atv
  */
-class RequestTest extends KernelTestBase {
-
+class RequestTest extends AtvKernelTestBase {
 
   /**
    * Request validator.
@@ -24,43 +22,17 @@ class RequestTest extends KernelTestBase {
    * @var \League\OpenAPIValidation\PSR7\RequestValidator
    */
   protected RequestValidator $validator;
-  /**
-   * The modules to load to run the test.
-   *
-   * @var array
-   */
-  protected static $modules = [
-      // Drupal.
-    'file',
-      // Contrib.
-    'openid_connect',
-      // Helfi modules.
-    'helfi_api_base',
-    'helfi_atv',
-    'helfi_atv_test',
-    'helfi_helsinki_profiili',
-      // Helsinki profiili requires audit log unnecessarily.
-    'helfi_audit_log',
-  ];
 
   /**
    * {@inheritdoc}
    */
   protected function setUp(): void {
     parent::setUp();
+
     $yamlFile = __DIR__ . '/Asiointitietovaranto.yaml';
     $this->validator = (new ValidatorBuilder)->fromYamlFile($yamlFile)->getRequestValidator();
 
-    $this->installConfig(['helfi_atv']);
-    putenv('ATV_API_KEY=fake');
-    putenv('ATV_USE_TOKEN_AUTH=true');
-    putenv('ATV_TOKEN_NAME=tokenName');
-    putenv('ATV_BASE_URL=127.0.0.1');
     putenv('ATV_VERSION=v1');
-    putenv('ATV_USE_CACHE=false');
-    putenv('APP_ENV=UNIT_TEST');
-    putenv('ATV_SERVICE=service');
-    putenv('ATV_MAX_PAGES=10');
   }
 
   /**
@@ -106,7 +78,7 @@ class RequestTest extends KernelTestBase {
       $match = $this->validator->validate($request);
     }
     catch (\Exception $e) {
-      $this->fail($e->getVerboseMessage());
+      $this->fail($e->getMessage());
       return;
     }
     $this->assertEquals('post', $match->method());
@@ -153,7 +125,7 @@ class RequestTest extends KernelTestBase {
       $match = $this->validator->validate($request);
     }
     catch (\Exception $e) {
-      $this->fail($e->getVerboseMessage());
+      $this->fail($e->getMessage());
       return;
     }
